+++ tests/v2/tasks_admin.spec.js
@@ -33,12 +33,10 @@ test('Tasks admin page [v2]', async ({ page, workflow }) => {
 		await waitPageLoading(page);
 	});
 
-	/** @type {number} */
-	let total;
 	await test.step('Search tasks without filters', async () => {
 		await searchTasks(page);
 		await expect(page.getByRole('table')).toBeVisible();
-		total = await page.getByRole('row').count();
+		const total = await page.getByRole('row').count();
 		expect(total).toBeGreaterThan(0);
 		await reset(page);
 	});
@@ -91,8 +89,6 @@ test('Tasks admin page [v2]', async ({ page, workflow }) => {
  * @param {import('@playwright/test').Page} page
  */
 async function searchTasks(page) {
-	// Increasing the results limit since during the tests many tasks may have been created
-	await page.getByRole('spinbutton', { name: 'Max number of results' }).fill('500');
 	await page.getByRole('button', { name: 'Search tasks' }).click();
 }